--- include/helperfunctions.h
@@ -1,8 +1,0 @@
-#ifndef __helperfunctions_H
-#define __helperfunctions_H
-
-/* Helper Functions used in the app that do not have a home */
-/* These should be refactored out over time into their respective libaries */
-
-
-#endif /*__ helperfunctions_H */
