+++ platformio.ini
@@ -19,7 +19,6 @@ lib_deps        =
     ;FatFs=https://github.com/stm32duino/FatFs.git@d6acb51
     SPI
 build_flags     =
-    -DARCH_STM32 ; Required for SPIMemory library
     -DBL_LEGACY_LEAF=
     -Wno-deprecated-declarations
     -Wno-comment
