--- src/helperfunctions.cpp
@@ -1,1 +1,0 @@
-#include "includes.h"
