+++ src/main.cpp
@@ -2,7 +2,6 @@
 //#include<SPIMemory.h>
 
 LCD lcd;
-EEPROM eeprom;
 URTouch  myTouch(PIN_SPI1_SCK , TOUCH_CS, PIN_SPI1_MOSI,PIN_SPI1_MISO, TOUCH_DI);
 EEPROMConfig eeprom_config;
 //SDCard sdcard;
@@ -22,7 +21,6 @@ void setup() {
   lcd.lcd_console_log("URTouch Init...");
   myTouch.InitTouch(1);
   myTouch.setPrecision(PREC_MEDIUM);
-  eeprom.init();
   eeprom_config.loadConfig();
   lcd.lcd_console_log("SD Test Init...");
   if(SD_Init() == 0){
