--- src/w25qxx.cpp
@@ -1,217 +1,0 @@
-#include "w25qxx.h"
-#include "SPI.h"
-
-
-#define W25Qxx_SPI     SPI
-
-void W25Qxx_SPI_CS_Set(uint8_t level)
-{
-  digitalWrite(FLASH_CS, level);
-}
-
-
-uint8_t W25Qxx_SPI_Read_Write_Byte(uint8_t data)
-{
-  return W25Qxx_SPI.transfer(data);
-}
-
-
-void W25Qxx_Init(void)
-{
-  //SPI1 (default): SPIFlash: Winbond 8MB (64Mb) Serial Flash Memory W25Q CS: PB9,MOSI=PA7,PIN_SPI_MISO=PA6,PIN_SPI_SCK=PA5
-  pinMode(FLASH_CS, OUTPUT);
-  W25Qxx_SPI.begin();
-  W25Qxx_SPI.setClockDivider(SPI_BAUDRATEPRESCALER_2);
-  W25Qxx_SPI_CS_Set(1);
-}
-
-
-void W25Qxx_WriteEnable(void)
-{
-  W25Qxx_SPI_CS_Set(0);
-  W25Qxx_SPI_Read_Write_Byte(CMD_WRITE_ENABLE);
-  W25Qxx_SPI_CS_Set(1);
-}
-
-void W25Qxx_WaitForWriteEnd(void)
-{
-  uint8_t flashstatus = 0;
-
-  W25Qxx_SPI_CS_Set(0);
-  W25Qxx_SPI_Read_Write_Byte(CMD_READ_REGISTER1);
-  do
-  {
-    flashstatus = W25Qxx_SPI_Read_Write_Byte(W25QXX_DUMMY_BYTE);
-  }
-  while ((flashstatus & 0x01) == SET); 
-  W25Qxx_SPI_CS_Set(1);
-}
-
-
-void W25Qxx_WritePage(uint8_t* pBuffer, uint32_t WriteAddr, uint16_t NumByteToWrite)
-{
-  W25Qxx_WriteEnable();
-  W25Qxx_SPI_CS_Set(0);
-	
-  W25Qxx_SPI_Read_Write_Byte(CMD_PAGE_PROGRAM);
-  W25Qxx_SPI_Read_Write_Byte((WriteAddr & 0xFF0000) >> 16);
-  W25Qxx_SPI_Read_Write_Byte((WriteAddr & 0xFF00) >> 8);
-  W25Qxx_SPI_Read_Write_Byte(WriteAddr & 0xFF);
-	
-  while (NumByteToWrite--)
-  {
-    W25Qxx_SPI_Read_Write_Byte(*pBuffer);
-    pBuffer++;
-  }
-	
-  W25Qxx_SPI_CS_Set(1);
-  W25Qxx_WaitForWriteEnd();
-}
-
-
-void W25Qxx_WriteBuffer(uint8_t* pBuffer, uint32_t WriteAddr, uint16_t NumByteToWrite)
-{
-  uint8_t NumOfPage = 0, NumOfSingle = 0, Addr = 0, count = 0, temp = 0;
-
-  Addr = WriteAddr % W25QXX_SPI_PAGESIZE;
-  count = W25QXX_SPI_PAGESIZE - Addr;
-  NumOfPage =  NumByteToWrite / W25QXX_SPI_PAGESIZE;
-  NumOfSingle = NumByteToWrite % W25QXX_SPI_PAGESIZE;
-
-  if (Addr == 0) /*!< WriteAddr is sFLASH_PAGESIZE aligned  */
-  {
-    if (NumOfPage == 0) /*!< NumByteToWrite < sFLASH_PAGESIZE */
-    {
-      W25Qxx_WritePage(pBuffer, WriteAddr, NumByteToWrite);
-    }
-    else /*!< NumByteToWrite > sFLASH_PAGESIZE */
-    {
-      while(NumOfPage--)
-      {
-        W25Qxx_WritePage(pBuffer, WriteAddr, W25QXX_SPI_PAGESIZE);
-        WriteAddr +=  W25QXX_SPI_PAGESIZE;
-        pBuffer += W25QXX_SPI_PAGESIZE;
-      }
-      W25Qxx_WritePage(pBuffer, WriteAddr, NumOfSingle);
-    }
-  }
-  else /*!< WriteAddr is not sFLASH_PAGESIZE aligned  */
-  {
-    if (NumOfPage == 0) /*!< NumByteToWrite < sFLASH_PAGESIZE */
-    {
-      if (NumOfSingle > count) /*!< (NumByteToWrite + WriteAddr) > sFLASH_PAGESIZE */
-      {
-        temp = NumOfSingle - count;
-
-        W25Qxx_WritePage(pBuffer, WriteAddr, count);
-        WriteAddr +=  count;
-        pBuffer += count;
-
-        W25Qxx_WritePage(pBuffer, WriteAddr, temp);
-      }
-      else
-      {
-        W25Qxx_WritePage(pBuffer, WriteAddr, NumByteToWrite);
-      }
-    }
-    else /*!< NumByteToWrite > sFLASH_PAGESIZE */
-    {
-      NumByteToWrite -= count;
-      NumOfPage =  NumByteToWrite / W25QXX_SPI_PAGESIZE;
-      NumOfSingle = NumByteToWrite % W25QXX_SPI_PAGESIZE;
-
-      W25Qxx_WritePage(pBuffer, WriteAddr, count);
-      WriteAddr +=  count;
-      pBuffer += count;
-
-      while (NumOfPage--)
-      {
-        W25Qxx_WritePage(pBuffer, WriteAddr, W25QXX_SPI_PAGESIZE);
-        WriteAddr +=  W25QXX_SPI_PAGESIZE;
-        pBuffer += W25QXX_SPI_PAGESIZE;
-      }
-
-      if (NumOfSingle != 0)
-      {
-        W25Qxx_WritePage(pBuffer, WriteAddr, NumOfSingle);
-      }
-    }
-  }
-}
-
-
-void W25Qxx_ReadBuffer(uint8_t* pBuffer, uint32_t ReadAddr, uint16_t NumByteToRead)
-{
-  W25Qxx_SPI_CS_Set(0);
-
-  W25Qxx_SPI_Read_Write_Byte(CMD_READ_DATA);
-
-  W25Qxx_SPI_Read_Write_Byte((ReadAddr & 0xFF0000) >> 16);
-  W25Qxx_SPI_Read_Write_Byte((ReadAddr& 0xFF00) >> 8);
-  W25Qxx_SPI_Read_Write_Byte(ReadAddr & 0xFF);
-
-  while (NumByteToRead--)
-  {
-    *pBuffer = W25Qxx_SPI_Read_Write_Byte(W25QXX_DUMMY_BYTE);
-    pBuffer++;
-  }
-	
-  W25Qxx_SPI_CS_Set(1);
-}
-
-uint32_t W25Qxx_ReadID(void)
-{
-  uint32_t Temp = 0;
-
-  W25Qxx_SPI_CS_Set(0);
-  W25Qxx_SPI_Read_Write_Byte(CMD_READ_ID);
-
-  Temp = W25Qxx_SPI_Read_Write_Byte(W25QXX_DUMMY_BYTE);
-  Temp = (Temp<<8) | W25Qxx_SPI_Read_Write_Byte(W25QXX_DUMMY_BYTE);
-  Temp = (Temp<<8) | W25Qxx_SPI_Read_Write_Byte(W25QXX_DUMMY_BYTE);
-
-  W25Qxx_SPI_CS_Set(1);
-
-  return Temp;
-}
-
-void W25Qxx_EraseSector(uint32_t SectorAddr)
-{
-  W25Qxx_WriteEnable();
-
-  W25Qxx_SPI_CS_Set(0);
-  W25Qxx_SPI_Read_Write_Byte(CMD_SECTOR_ERASE);
-  W25Qxx_SPI_Read_Write_Byte((SectorAddr & 0xFF0000) >> 16);
-  W25Qxx_SPI_Read_Write_Byte((SectorAddr & 0xFF00) >> 8);
-  W25Qxx_SPI_Read_Write_Byte(SectorAddr & 0xFF);
-  W25Qxx_SPI_CS_Set(1);
-
-  W25Qxx_WaitForWriteEnd();
-}
-
-void W25Qxx_EraseBlock(uint32_t BlockAddr)
-{
-  W25Qxx_WriteEnable();
-	
-  W25Qxx_SPI_CS_Set(0);
-  W25Qxx_SPI_Read_Write_Byte(CMD_BLOCK_ERASE);
-  W25Qxx_SPI_Read_Write_Byte((BlockAddr & 0xFF0000) >> 16);
-  W25Qxx_SPI_Read_Write_Byte((BlockAddr & 0xFF00) >> 8);
-  W25Qxx_SPI_Read_Write_Byte(BlockAddr & 0xFF);
-  W25Qxx_SPI_CS_Set(1);
-
-  W25Qxx_WaitForWriteEnd();
-}
-
-void W25Qxx_EraseBulk(void)
-{
-  W25Qxx_WriteEnable();
-
-  W25Qxx_SPI_CS_Set(0);
-  W25Qxx_SPI_Read_Write_Byte(CMD_FLASH__BE);
-  W25Qxx_SPI_CS_Set(1);
-
-  W25Qxx_WaitForWriteEnd();
-}
-
-
